Move the ROI patching function to a new utility module
